迁移为mocha测试环境
1. 原来的测试是用test.bat触发的，在Win之外的其他平台不适用。
2. 搭建基础的测试环境，现在可以用命令
 npm test 触发递归测试
