+++ test/getExcelArrCb_test.js
@@ -1,11 +1,21 @@
-var ejsExcel = require("../ejsExcel");
 var fs = require("fs");
-//获得Excel模板的buffer对象
-var exlBuf = fs.readFileSync("./test.xlsx");
+var path=require('path');
+var ejsExcel = require("../ejsExcel");
 
-//getExcelArr 返回Promise对象
-ejsExcel.getExcelArr(exlBuf).then(function(exlJson) {
-	console.log(exlJson);
-}).catch(function(err){
-	console.error(err);
+
+var EXCEL_PATH=path.join(__dirname,"./test.xlsx");
+describe('test get Excel Attribute', function () {
+
+    it('test test.xlsx',function () {
+        //获得Excel模板的buffer对象
+        var exlBuf = fs.readFileSync(EXCEL_PATH);
+
+        //getExcelArr 返回Promise对象
+        return ejsExcel.getExcelArr(exlBuf)
+            .then(function(exlJson) {
+                console.log(exlJson);
+            }).catch(function(err){
+                console.error(err);
+            });
+    });
 });
