--- test/test.bat
@@ -1,5 +1,0 @@
-%~d0
-cd %~dp0
-node testExcel.js
-node testExcel2.js
-pause
